Add add test
