+++ package.json
@@ -1,38 +1,41 @@
 {
-  "name": "todolist",
+  "name": "webpack-exercise",
   "version": "1.0.0",
-  "description": "![68747470733a2f2f696d672e796f75747562652e636f6d2f76692f416355642d5f596a6a71672f302e6a7067](https://user-images.githubusercontent.com/106747387/230112602-9404e3e1-7fbb-4690-98e3-bbffbf5c314a.jpeg)",
+  "description": "",
   "private": true,
   "scripts": {
     "test": "jest",
-    "build": "webpack",
-    "start": "webpack serve"
+    "watch": "webpack --watch",
+    "start": "webpack serve --open",
+    "build": "webpack"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "@babel/plugin-transform-modules-commonjs": "^7.22.5",
+    "@babel/plugin-transform-modules-commonjs": "^7.21.2",
     "babel-eslint": "^10.1.0",
-    "css-loader": "^6.8.1",
+    "css-loader": "^6.7.3",
     "eslint": "^7.32.0",
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-plugin-import": "^2.27.5",
-    "hint": "^7.1.8",
-    "html-webpack-plugin": "^5.5.1",
+    "hint": "^7.1.7",
+    "html-webpack-plugin": "^5.5.0",
     "jest": "^29.5.0",
     "jest-environment-jsdom": "^29.5.0",
-    "style-loader": "^3.3.3",
+    "jest-environment-jsdom-global": "^4.0.0",
+    "jest-localstorage-mock": "^2.4.26",
+    "jsdom": "^21.1.1",
+    "style-loader": "^3.3.2",
     "stylelint": "^13.13.1",
     "stylelint-config-standard": "^21.0.0",
     "stylelint-csstree-validator": "^1.9.0",
     "stylelint-scss": "^3.21.0",
-    "webpack": "^5.85.1",
-    "webpack-cli": "^5.1.3",
-    "webpack-dev-server": "^4.15.0"
+    "webpack": "^5.76.3",
+    "webpack-cli": "^5.0.1",
+    "webpack-dev-server": "^4.13.1"
   },
   "dependencies": {
     "lodash": "^4.17.21"
-  },
-  "main": "webpack.config.js"
+  }
 }
