+++ src/add.test.js
@@ -4,23 +4,17 @@
 
 import './__mocks__/mockHtml.js';
 import { AddToDo } from './add.js';
-
+import ToDo from './render.js';
 
 describe('test add ', () => {
   it('Add tasks to local storage', () => {
-    
-    const input = document.querySelectorAll('list-container');
-    console.log(new AddToDo());
     const newAddToDo = new AddToDo();
 
-    newAddToDo.addToDo('hello')
-    newAddToDo.addToDo('start')
-    newAddToDo.addToDo('middle')
-    newAddToDo.addToDo('finish')
-    newAddToDo.addToDo('well-done')
-
-    // const ssss = JSON.parse(localStorage.getItem('toDo'))
-    // console.log(ssss)
+    newAddToDo.addToDo('hello');
+    newAddToDo.addToDo('start');
+    newAddToDo.addToDo('middle');
+    newAddToDo.addToDo('finish');
+    newAddToDo.addToDo('well-done');
 
     expect(JSON.parse(localStorage.getItem('toDo'))).toEqual([
       { description: 'hello', completed: false, num: 1 },
@@ -29,5 +23,20 @@ describe('test add ', () => {
       { description: 'finish', completed: false, num: 4 },
       { description: 'well-done', completed: false, num: 5 },
     ]);
+  });
+
+  it('Add tasks one li to html', () => {
+    const newAddToDo = new AddToDo();
+    newAddToDo.addToDo('wel-done');
+    const newrender = new ToDo();
+    newrender.renderToDo();
+    const list = document.querySelectorAll('.wrapper');
+    const initiallength = list.length;
+    newAddToDo.addToDo('wel-done');
+    const newrender1 = new ToDo();
+    newrender1.renderToDo();
+    const finallength = document.querySelectorAll('.wrapper').length;
+
+    expect(finallength).toEqual(initiallength + 1);
   });
 });
